* Several bugfixes in the ctl::Dot controller related to handling logarithmic values.
* Added possibility to evaluate substituted text values.
* Introduced ctl::LineSegment widget controller.
* Fixed memory leakage related to not complete release of allocated memory on UI destruction.
* Added generation of explicit element definitions for LV2 groups in TTL files.
* Added flags to ui::IPortListener::notify() method that allow to indicate the source of the event.
* Updated module versions in dependencies.

+++ include/lsp-plug.in/plug-fw/wrap/lv2/ports.h
@@ -464,7 +464,7 @@ namespace lsp
                      value       = meta::limit_value(pMetadata, value);
                      if (pMetadata->flags & meta::F_PEAK)
                      {
-                         if (fabs(fValue) < fabs(value))
+                         if (fabsf(fValue) < fabsf(value))
                              fValue = value;
                      }
                      else
